pyproject.toml
[project]
name = "ai-coding-agent-langchain-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "cryptography>=43.0.0",
    "docker>=7.1.0",
    "flask>=3.1.2",
    "flask-apscheduler>=1.13.1",
    "flask-sqlalchemy>=3.1.1",
    "grandalf>=0.8",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=3.2.0",
    "langchain-mistralai>=0.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.1",
    "mcp>=1.22.0",
    "mcp-server-git>=2025.9.25",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
